chore: minor structure improvements
