+++ apps/frontend-manage/src/components/analytics/activity/TotalStudentActivityPlot.tsx
@@ -14,6 +14,7 @@ import {
   XAxis,
   YAxis,
 } from 'recharts'
+import ActivityLevelTag from './ActivityLevelTag'
 import useTotalStudentActivityHistogram from './useTotalStudentActivityHistogram'
 
 interface TotalStudentActivityPlotProps {
@@ -185,30 +186,6 @@ function TotalStudentActivityPlot({
       </div>
     </div>
   )
-}
-
-const ActivityLevelTag = ({ activityLevel }: { activityLevel: 1 | 2 | 3 }) => {
-  const t = useTranslations()
-
-  if (activityLevel === 3) {
-    return (
-      <div className="w-max rounded-md border bg-green-700 px-1 py-0.5 text-white">
-        {t('manage.analytics.activityLevelHigh')}
-      </div>
-    )
-  } else if (activityLevel === 2) {
-    return (
-      <div className="w-max rounded-md border bg-yellow-500 px-1 py-0.5 text-white">
-        {t('manage.analytics.activityLevelMedium')}
-      </div>
-    )
-  } else {
-    return (
-      <div className="w-max rounded-md border bg-red-600 px-1 py-0.5 text-white">
-        {t('manage.analytics.activityLevelLow')}
-      </div>
-    )
-  }
 }
 
 export default TotalStudentActivityPlot
