+++ packages/i18n/messages/de.ts
@@ -1856,7 +1856,7 @@ Da die KlickerUZH-App noch nicht im iOS-App-Store verfügbar ist, folgen Sie die
       activityLevel: 'Aktivitätslevel',
       activityLevelHigh: 'HOCH',
       activityLevelMedium: 'MITTEL',
-      activityLevelLow: 'NIEDRIG',
+      activityLevelLow: 'TIEF',
     },
   },
   control: {
